requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0  # For OpenAI client compatibility with Ollama
pydantic>=2.0.0
jsonschema>=4.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
ollama>=0.4.7